Add migration-Rename entity ProductInTheCart to Cart, remove UserId & DiscountCodeId FKs from OrderProduct entity, Rename properties in Product & Cart, Add missing ICollection inverse properties. Change some ViewModels. Continue shopping is using javascript.window.history.back() etc

+++ src/Web/TizianaTerenzi.Web.ViewModels/Orders/OrdersListingViewModel.cs
@@ -16,7 +16,7 @@
 
         public string DiscountCodeName { get; set; }
 
-        public double? DiscountCodeDiscount { get; set; }
+        public byte? DiscountCodeDiscount { get; set; }
 
         public DateTime CreatedOn { get; set; }
